src/main/resources/scripts/requirements.txt
geopandas>=0.10.2
shapely>=1.7.1
numpy>=1.21.2
pandas>=1.3.3
filterpy>=1.4.5
matplotlib>=3.4.3
psycopg2-binary>=2.9.1
